Move jPathToCPath to CPathUtils
