+++ yggdrasil/src/main/scala/quasar/yggdrasil/table/Slice.scala
@@ -1718,7 +1718,7 @@ abstract class Slice { source =>
     columns.foldLeft[RValue](CUndefined) {
       case (rv, (ColumnRef(selector, _), col)) if col.isDefinedAt(row) =>
         CPathUtils.cPathToJPaths(selector, col.cValue(row)).foldLeft(rv) {
-          case (rv, (path, value)) => rv.unsafeInsert(jPathToCPath(path), value)
+          case (rv, (path, value)) => rv.unsafeInsert(CPathUtils.jPathToCPath(path), value)
         }
 
       case (rv, _) => rv
@@ -2108,7 +2108,7 @@ object Slice {
       case (acc, (jpath, JUndefined)) => acc
       case (acc, (jpath, v)) =>
         val ctype = CType.forJValueRaw(v) getOrElse { sys.error("Cannot determine ctype for " + v + " at " + jpath + " in " + jv) }
-        val ref   = ColumnRef(remapPath.map(_ (jpath)).getOrElse(jPathToCPath(jpath)), ctype)
+        val ref   = ColumnRef(remapPath.map(_ (jpath)).getOrElse(CPathUtils.jPathToCPath(jpath)), ctype)
 
         val updatedColumn: ArrayColumn[_] = v match {
           case JBool(b) =>
