+++ yggdrasil/src/main/scala/quasar/yggdrasil/util/CPathUtils.scala
@@ -21,6 +21,16 @@ import quasar.blueeyes.json._
 import quasar.precog.common._
 
 object CPathUtils {
+
+  def jPathToCPath(jpath: JPath): CPath = {
+    val nodes2 = jpath.nodes map {
+      case JPathField(name) => CPathField(name)
+      case JPathIndex(idx)  => CPathIndex(idx)
+    }
+
+    CPath(nodes2: _*)
+  }
+
   def cPathToJPaths(cpath: CPath, value: CValue): List[(JPath, CValue)] = (cpath.nodes, value) match {
     case (CPathField(name) :: tail, _) => addComponent(JPathField(name), cPathToJPaths(CPath(tail), value))
     case (CPathIndex(i) :: tail, _)    => addComponent(JPathIndex(i), cPathToJPaths(CPath(tail), value))
