+++ yggdrasil/src/test/scala/quasar/yggdrasil/table/BlockLoadSpec.scala
@@ -22,6 +22,7 @@ import quasar.pkg.tests._, Gen._
 import quasar.precog.common._
 import SampleData._
 import SJValueGenerators._
+import quasar.yggdrasil.util.CPathUtils
 
 import scalaz._
 import scalaz.syntax.std.boolean._
@@ -79,7 +80,7 @@ trait BlockLoadSpec extends SpecificationLike with ScalaCheck {
       (back \ "value" != JUndefined).option(back)
     }
 
-    val cschema = module.schema map { case (jpath, ctype) => ColumnRef(jPathToCPath(jpath), ctype) }
+    val cschema = module.schema map { case (jpath, ctype) => ColumnRef(CPathUtils.jPathToCPath(jpath), ctype) }
 
     val result = module.Table.constString(Set("/test")).load(Schema.mkType(cschema).get).flatMap(t => EitherT.rightT(t.toJson)).run.unsafeRunSync
     result.map(_.toList) must_== \/.right(expected.toList.map(RValue.fromJValueRaw))
